backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
SQLAlchemy
python-docx
ffmpeg-python
faster-whisper
webrtcvad
python-magic
gunicorn
